Allow creating a tag without a message.
gitg only creates annotated or signed tags. However, sometimes you
just don't want to use a message in your tag. This patch enables you
to do that.

If you give a message, but don't ask to sign the tag, the tag will be
an annotated tag. If you give a message and ask to sign the tag, the
tag will be a signed tag (zomg). If you omit a message, then the tag
will be created without a message (not signed or annotated).

The UI will complain appropriately if you screw up.

Signed-off-by: Jonny Lamb <jonnylamb@gnome.org>

https://bugzilla.gnome.org/show_bug.cgi?id=611434