Improve keynav some more
The previous commit made it so that arrow up/down can leave the
list box when the last child is reached. But it did not work for
Page up/down, due to the way the code was written. With this
commit, all up/down key bindings can leave the widget.