Add hack to fix parallel install
libtool needs to relink at install (on some platforms) which requires
dependency libraries to be already installed. automake does not
generate the appropriate dependency so we do it here automatically. As
an additional bonus, we have to trick automake using a variable
indirection for the rule because it would otherwise omit the default
install rule.