Open selected folder first when opening repository
When opening a repository gitg takes the current folder from the file
chooser and attempts to open it as a repository. The user may however
select a folder in the file chooser, which may be the actual repository.
Attempt to open the selected folder and only fall back to the current
folder if nothing is selected.