Select ref rows on button press
ListBox is designed to handle gestures, which means that it will wait
for a release to actually select a row. This in turn means that there
is a user noticeable delay in selecting rows in the refs list with the
mouse. We override button_press_event to select the row directly on a
button press.