Fix window growing each start
As described in bug 771290, the window got slightly larger each time
gitg was run.  This was due to the way we accidentally mixed APIs when
saving/restoring the window size.

To fix this, use GtkWindow methods when saving the window size.  This
is consistent with our prior use of GtkWindow methods to set the window
size on startup.

https://bugzilla.gnome.org/show_bug.cgi?id=771290