Allow popping menu up with keyboard
Currently, the code assumes that any attempt to pop up the menu will
come from a button press event, which is not true when using the
keyboard, in which case the assumption causes a crash.

This commit adds the ability to pop up the menu using the keyboard as
well.

https://bugzilla.gnome.org/show_bug.cgi?id=782275