build: Change project's default values
The project license has a different license from the one in the
source code files: GPL2 in the project's description, and GPL2 or
later in the source code files.

The default options for warning level and the no undefined flag when
linking have also the default options already used by meson.

This patch fixes the project license and also removes the default
options which are already used by meson.

https://bugzilla.gnome.org/show_bug.cgi?id=788796