build: Fix shared library paths in typelib files
The GObject Introspection metadata generation by meson uses a custom
target command[0]. This command also needs to specify the shared
library where the symbols in the typelib file can be found.

However, the values used are wrong because the path to the built
shared libraries are specified instead of the name of the
shared libraries themselves.

This has been fixed by specifying the name of the shared libraries.

Fixes #157

[0] http://mesonbuild.com/Vala.html#gobject-introspection