Make the Commit dialog window bigger by default
Provide sufficient width to display the right margin column,
even with the "Large Text" (1.25 font scaling) accessibility setting.

Also increase height, so that longer commit messages can be drafted,
without the user needing to resize the window every time.

As this only sets the default size, it is compatible with small screens;
the window manager can size it down if space is constrained.

Fixes issue #233